HAPP-1904 - updated login error text - updated target SDK version to 34 - updated play core SDK version to support android 34 - fixed issue of quick link tile after deleting all data from app.
